requirements.txt
openvino-nightly
numpy
transformers==4.38.0
jinja2
